+++ src/test/java/blackjack/model/player/matcher/MatcherTest.java
@@ -4,14 +4,12 @@ import static blackjack.model.card.Suit.CLOVER;
 import static blackjack.model.card.Suit.DIAMOND;
 import static blackjack.model.card.Suit.HEART;
 import static blackjack.model.card.Suit.SPADE;
-import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.model.card.Card;
 import blackjack.model.card.Rank;
 import blackjack.model.player.Dealer;
 import blackjack.model.player.Gamer;
-import blackjack.model.player.Money;
 import java.math.BigDecimal;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
