+++ src/test/java/blackjack/model/player/matcher/MoneyTest.java
@@ -1,6 +1,5 @@
-package blackjack.model.player;
+package blackjack.model.player.matcher;
 
-import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.math.BigDecimal;
