+++ src/test/java/blackjack/model/player/matcher/ResultTest.java
@@ -2,7 +2,6 @@ package blackjack.model.player.matcher;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.model.player.Money;
 import java.math.BigDecimal;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
